+++ src/main/java/com/chhin/fitnesstracker/entities/BodyMeasurementArea.java
@@ -12,16 +12,16 @@ public class BodyMeasurementArea {
   @SequenceGenerator(name = "body_measurement_area_gen", sequenceName = "body_measurement_area_pk_seq", allocationSize = 1)
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "body_measurement_area_gen")
   @Column(name = "body_measurement_area_id")
-  private Integer bodyMeasurementAreaId;
+  private Long bodyMeasurementAreaId;
 
   @Column(name = "body_measurement_area_description", nullable = false, unique = true, length = 100)
   private String bodyMeasurementAreaDescription;
 
-  public Integer getBodyMeasurementAreaId() {
+  public Long getBodyMeasurementAreaId() {
     return bodyMeasurementAreaId;
   }
 
-  public void setBodyMeasurementAreaId(Integer bodyMeasurementAreaId) {
+  public void setBodyMeasurementAreaId(Long bodyMeasurementAreaId) {
     this.bodyMeasurementAreaId = bodyMeasurementAreaId;
   }
