+++ src/main/java/com/chhin/fitnesstracker/service/LoggedInUserService.java
@@ -5,10 +5,11 @@ import com.chhin.fitnesstracker.entities.FTUser;
 import java.util.Optional;
 
 public class LoggedInUserService {
-  private final int userId;
+
+  private final Long userId;
   private final FTUserService ftUserService;
 
-  public LoggedInUserService(int userId, FTUserService ftUserService) {
+  public LoggedInUserService(Long userId, FTUserService ftUserService) {
     this.userId = userId;
     this.ftUserService = ftUserService;
   }
